+++ app-modules/publishing/src/Models/PostVersions.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Domains\Publishing\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class PostVersions extends Model
+{
+    //
+}